Move MKCoordinateRegion extension to Helper
