+++ TransEuroTrail/Shared/TrackDetailView.swift
@@ -15,37 +15,6 @@ struct Location: Identifiable {
     let coordinate: CLLocationCoordinate2D
 }
 
-extension MKCoordinateRegion {
-    init?(gpxWaypoints: [GPXWaypoint]) {
-        var minLat: CLLocationDegrees = 90.0
-        var maxLat: CLLocationDegrees = -90.0
-        var minLon: CLLocationDegrees = 180.0
-        var maxLon: CLLocationDegrees = -180.0
-        
-        for waypoint in gpxWaypoints {
-            guard let lat = waypoint.latitude, let long = waypoint.longitude else {
-                return nil
-            }
-            if lat < minLat {
-                minLat = lat
-            }
-            if long < minLon {
-                minLon = long
-            }
-            if lat > maxLat {
-                maxLat = lat
-            }
-            if long > maxLon {
-                maxLon = long
-            }
-        }
-        
-        let span = MKCoordinateSpan(latitudeDelta: (maxLat - minLat)*2.0, longitudeDelta: (maxLon - minLon)*2.0)
-        let center = CLLocationCoordinate2DMake(maxLat - span.latitudeDelta / 4, maxLon - span.longitudeDelta / 4)
-        self.init(center: center, span: span)
-    }
-}
-
 struct TrackDetailView: View {
     private let gpx = GPXParser(withData: Bundle.main.loadData(from: "Germany.gpx")).parsedData()
